perf(web): split ChatPage into independent re-render subtrees
ChatPage's root component subscribed to the streaming-hot store fields
(blocks, activeResponse, pendingUserMessages, interruptedResponseIds) and
built the bubble list from them, so every SSE frame (~60/s during
streaming) re-rendered the whole root: it recomputed all the liveness /
permission / model-option derivations and reconciled MainAgentSurface, the
Composer, and every side surface — none of which depend on blocks.

Move the streaming-hot subscriptions and the entire bubble pipeline into a
new memoized <Transcript> subtree (the only thing that re-renders per
frame). The root now derives hasPendingElicitation through an edge-stable
boolean selector instead of subscribing to the blocks array, so it wakes
only on turn / elicitation edges. MainAgentSurface no longer receives or
builds bubbles; it renders <Transcript> alongside the composer and chrome,
which bail out of streaming-frame re-renders.

To keep <Transcript> free of an import cycle with ChatPage, the shared
bubble-rendering components, scroll helpers, and pure bubble functions move
to web/src/components/chat/chatBubbleParts.tsx; ChatPage re-exports them so
existing importers (tests + components) keep resolving them via ./ChatPage.

No behavior change — pure decomposition.
